+++ src/test/kotlin/org/taktik/freehealth/middleware/service/impl/AgreementServiceUtilsTest.kt
@@ -15,12 +15,13 @@ import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.junit4.SpringRunner
 import org.taktik.connector.technical.utils.ConnectorXmlUtils
 import org.taktik.connector.technical.utils.MarshallerHelper
+import org.taktik.freehealth.middleware.service.AgreementServiceUtils
 import org.taktik.icure.fhir.entities.r4.practitionerrole.PractitionerRole
 import org.taktik.icure.fhir.entities.r4.reference.Reference
 
 class AgreementServiceUtilsTest {
 
-    val agreementServiceUtils = AgreementServiceUtils()
+    val agreementServiceUtils = AgreementServiceUtilsImpl()
     @Test
     fun getPractitionerRole() {
         val role: String = "persphysiotherapist"
@@ -75,7 +76,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getCodableConcept() {
-        val agreementServiceUtils = AgreementServiceUtils()
 
         val codeableConcept = agreementServiceUtils.getCodableConcept("https://example.com/system", "exampleCode")
 
@@ -97,7 +97,6 @@ class AgreementServiceUtilsTest {
     @Test
     fun getClaim() {
         // Créer une instance de la classe AgreementServiceUtils
-        val agreementServiceUtils = AgreementServiceUtils()
 
         // Appeler la fonction que vous souhaitez tester
         val claim = agreementServiceUtils.getClaim(
@@ -124,7 +123,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getSupportingInfo() {
-        val agreementServiceUtils = AgreementServiceUtils()
 
         val supportingInfo1 = agreementServiceUtils.getSupportingInfo(1, "attachment", "functional-report", null, null, "QW5uZXhlIGlubGluZSwgYmFzZTY0ZWQ=", "nom/description de l'annexe", "application/pdf")
         val supportingInfo2 = agreementServiceUtils.getSupportingInfo(2, "info", null, null, "additional Information", null, null, null)
@@ -160,7 +158,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getOrganization() {
-        val agreementServiceUtils = AgreementServiceUtils()
         val organizationId = "1";
         val orgNihii = "71000436000"
         val orgType = "orghospital"
@@ -185,7 +182,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getPatient() {
-        val agreementServiceUtils = AgreementServiceUtils()
 
         val patientWithSsin = agreementServiceUtils.getPatient("Jean", "Dupont", "male", "73031805784", null, null)
         val patientWithoutSsin = agreementServiceUtils.getPatient("Jean", "Dupont", "male", null, "109", "45613414615SDE")
@@ -222,7 +218,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getServiceRequest() {
-        val agreementServiceUtils = AgreementServiceUtils()
 
         val serviceRequest = agreementServiceUtils.getServiceRequest(
             "1",
@@ -266,7 +261,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getContained() {
-        val agreementServiceUtils = AgreementServiceUtils()
 
         val containedList = agreementServiceUtils.getContained("pdfData", "123")
 
@@ -284,7 +278,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getParameters() {
-        val agreementServiceUtils = AgreementServiceUtils()
 
         val parameterId = "1"
         val parameterNames = arrayOf("resourceType", "patient", "use", "subType", "preAuthPeriod")
@@ -321,7 +314,6 @@ class AgreementServiceUtilsTest {
 
     @Test
     fun getParameter() {
-        val agreementServiceUtils = AgreementServiceUtils()
 
         val parameterName = "subType"
         val agreementTypes = "someAgreementType"
